fixed bug when creating folder

+++ icon_fetcher.py
@@ -5,7 +5,8 @@ import os
 import shutil
 
 new_folder = "icons"
-shutil.rmtree(new_folder)
+if os.path.exists(new_folder):
+    shutil.rmtree(new_folder)
 os.makedirs(new_folder)
 
 URL_IMG = f"https://raw.communitydragon.org/latest/plugins/rcp-be-lol-game-data/global/default/v1/champion-icons"
